feat: completed manager

+++ src/downloader/util.rs
@@ -2,7 +2,7 @@ use reqwest::{Client, Request};
 use url::Url;
 use crate::downloader::error::DownloadError;
 
-pub async fn get_file_length(url: Url) -> anyhow::Result<Option<usize>, DownloadError> {
+pub async fn get_file_length(url: Url) -> anyhow::Result<Option<u64>, DownloadError> {
     let request = Request::new(reqwest::Method::HEAD, url);
     let client = Client::new();
 
@@ -10,7 +10,7 @@ pub async fn get_file_length(url: Url) -> anyhow::Result<Option<usize>, Download
         Ok(response) => {
             if let Some(content_length) = response.headers().get(reqwest::header::CONTENT_LENGTH) {
                 if let Ok(length) = content_length.to_str() {
-                    if let Ok(length) = length.parse::<usize>() {
+                    if let Ok(length) = length.parse::<u64>() {
                         return Ok(Some(length))
                     }
                 }
